backend/pyproject.toml
[project]
name = "pickle-reef-backend"
version = "0.1.0"
description = "FastAPI backend for the Pickle Reef aquarium controller"
authors = [{ name = "Pickle Reef" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi~=0.115",
    "uvicorn[standard]~=0.30",
    "sqlmodel~=0.0.21",
    "pydantic-settings~=2.6",
    "python-dotenv~=1.0",
    "aiosqlite~=0.20",
    "paho-mqtt~=2.1",
    "alembic~=1.13"
]

[project.optional-dependencies]
dev = [
    "pytest~=8.2",
    "pytest-asyncio~=0.23",
    "httpx~=0.27",
    "ruff~=0.6"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
